+++ RidePal/src/main/java/com/example/ridepal/exceptions/EntityNotFoundException.java
@@ -1,7 +1,7 @@
 package com.example.ridepal.exceptions;
 
 public class EntityNotFoundException extends RuntimeException{
-    public EntityNotFoundException(String type, int id) {
+    public EntityNotFoundException(String type, long id) {
         this(type, "id", String.valueOf(id));
     }
