+++ RidePal/src/main/java/com/example/ridepal/repositories/ArtistRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistRepository extends JpaRepository<Artist, Integer> {
 
-    Artist findById(int id);
+    Artist findById(long id);
 }
